Moved login-buttons into ui components
